+++ src/main/java/com/minecraftabnormals/autumnity/common/block/FoulBerryBushBlock.java
@@ -83,7 +83,7 @@ public class FoulBerryBushBlock extends BushBlock implements IGrowable {
 	public void entityInside(BlockState state, World worldIn, BlockPos pos, Entity entityIn) {
 		if (entityIn instanceof LivingEntity && entityIn.getType() != EntityType.BEE && entityIn.getType() != AutumnityEntities.SNAIL.get() && entityIn.getType() != AutumnityEntities.TURKEY.get()) {
 			LivingEntity livingentity = ((LivingEntity) entityIn);
-			entityIn.makeStuckInBlock(state, new Vector3d((double) 0.8F, 0.75D, (double) 0.8F));
+			entityIn.makeStuckInBlock(state, new Vector3d(0.8F, 0.75D, 0.8F));
 			if (!worldIn.isClientSide && !livingentity.hasEffect(Effects.POISON) && !livingentity.isShiftKeyDown()) {
 				livingentity.addEffect(new EffectInstance(Effects.POISON, 120));
 			}
