+++ src/main/java/com/minecraftabnormals/autumnity/common/block/MapleLeavesBlock.java
@@ -23,15 +23,15 @@ public class MapleLeavesBlock extends AbnormalsLeavesBlock {
 			BlockPos blockpos = pos.below();
 			if (worldIn.isEmptyBlock(blockpos)) {
 				int color = worldIn.getBiome(pos).getFoliageColor();
-				
+
 				double d0 = (color >> 16 & 255) / 255.0F;
 				double d1 = (color >> 8 & 255) / 255.0F;
 				double d2 = (color & 255) / 255.0F;
-				
-				double d3 = (double) ((float) pos.getX() + rand.nextFloat());
+
+				double d3 = (float) pos.getX() + rand.nextFloat();
 				double d4 = (double) pos.getY() - 0.05D;
-				double d6 = (double) ((float) pos.getZ() + rand.nextFloat());
-				
+				double d6 = (float) pos.getZ() + rand.nextFloat();
+
 				worldIn.addParticle(AutumnityParticles.FALLING_MAPLE_LEAF.get(), d3, d4, d6, d0, d1, d2);
 			}
 		}
